requirements.txt
affine==2.4.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
filelock==3.13.1
fsspec==2023.12.0
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
Pillow==10.1.0
pyparsing==3.1.1
rasterio==1.3.9
requests==2.31.0
snuggs==1.4.7
sympy==1.12
torch==2.1.1
torchvision==0.16.1
typing_extensions==4.8.0
urllib3==2.1.0
